Cut halt

+++ src/instrs.rs
@@ -1,7 +1,7 @@
 use std::collections::HashMap;
 
 pub type Color = u64;
-pub type State = i32;
+pub type State = u64;
 pub type Shift = bool;
 
 pub type Slot = (State, Color);
